requirements.txt
streamlit
pdfplumber
pandas
numpy
python-dotenv
plotly
google-genai
